requirements.txt
alembic==1.5.4
click==7.1.2
fastapi==0.63.0
flake8==3.8.4
h11==0.12.0
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
passlib==1.7.4
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pydantic==1.7.3
pyflakes==2.2.0
PyJWT==2.0.1
python-dateutil==2.8.1
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.23
starlette==0.13.6
uvicorn==0.13.3
